requirements.txt
pysrt
webvtt-py
argparse